requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
faiss-cpu==1.8.0
filelock==3.13.1
fsspec==2024.3.1
ftfy==6.2.0
huggingface-hub==0.21.4
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opensearch-py==2.4.2
packaging==24.0
pandas==2.2.1
pillow==10.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
sentencepiece==0.2.0
six==1.16.0
sympy==1.12
tokenizers==0.15.2
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
